#903 refactor unit tests - make setup more scalable
