+++ Consensust.cpp
@@ -57,38 +57,7 @@
 #include <gperftools/heap-profiler.h>
 #endif
 
-
-ConsensusEngine* engine;
-
-
-class DontCleanup {
-public:
-    DontCleanup() { Consensust::setConfigDirPath( boost::filesystem::system_complete( "." ) ); };
-
-    ~DontCleanup() {}
-};
-
-
-class StartFromScratch {
-public:
-    StartFromScratch() {
-        int i = system( "rm -rf /tmp/*.db.*" );
-        i = system( "rm -rf /tmp/*.db" );
-        i++;  // make compiler happy
-        Consensust::setConfigDirPath( boost::filesystem::system_complete( "." ) );
-
-#ifdef GOOGLE_PROFILE
-        HeapProfilerStart( "/tmp/consensusd.profile" );
-        HeapProfilerStart( "/tmp/consensusd.profile" );
-#endif
-    };
-
-    ~StartFromScratch() {
-#ifdef GOOGLE_PROFILE
-        HeapProfilerStop();
-#endif
-    }
-};
+ConsensusEngine* engine;  // definition
 
 uint64_t Consensust::getRunningTimeS() {
     if ( runningTimeS == 0 ) {
@@ -132,7 +101,7 @@ void abort_handler( int ) {
     exit( 0 );
 }
 
-block_id basicRun( int64_t _lastId = 0 ) {
+block_id basicRun( int64_t _lastId ) {
     try {
         REQUIRE( ConsensusEngine::getEngineVersion().size() > 0 );
 
@@ -191,7 +160,3 @@ void exit_check() {
         usleep( 100 * 1000 );
     }
 }
-
-
-#include "unittests/consensus_tests.cpp"
-#include "unittests/sgx_tests.cpp"
